Рефакторинг. Вынес методы в отдельные классы
